farmer/requirements.txt
pandas
numpy
matplotlib
joblib
scikit-learn
Pillow 
torch
torchvision
torchsummary